+++ duplicity/diffdir.py
@@ -736,5 +736,5 @@ def get_block_size(file_len):
         return 512 # set minimum of 512 bytes
     else:
         # Split file into about 2000 pieces, rounding to 512
-        file_blocksize = long((file_len / (2000 * 512)) * 512)
+        file_blocksize = int((file_len / (2000 * 512)) * 512)
         return min(file_blocksize, globals.max_blocksize)
